partially grabbable and socket
i only added grabbable to all object that have name contain "computer" and "chair", coffee cup, and etc ( i forget le). Socket i only add on the table in first and second floor name "SocketTable").

I not yet add in the item that we choose to grab to put on socket.

Free to modify until u feel better when using